requirements.txt
alembic==1.16.2
aniso8601==10.0.1
attrs==25.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
rpds-py==0.25.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
